requirements.txt
Django>=1.9
django-easy-pdf==0.1.0
psycopg2==2.6.2
git+https://github.com/chrisglass/xhtml2pdf.git